feat: changed to server components
