+++ src/components/Blocks/TextImage/TextImage.tsx
@@ -1,17 +1,15 @@
-'use client';
-
 import React from 'react';
-import { Image, StructuredText } from 'react-datocms';
+import { SRCImage, StructuredText } from 'react-datocms';
 import ImageBlock from '@Blocks/Image/Image';
 import MainHeading from 'components/Blocks/MainHeading/MainHeading';
-import { motion } from 'framer-motion';
 import type {
-  FileField,
   ImageRecord,
   MainHeadingRecord,
   TextImageRecord,
 } from 'infrastructure/generated/graphql';
 import { cn } from 'lib/utils';
+
+import TextImageAnimation from './TextImageAnimation';
 
 export default function TextImageBlock({
   imageLocation,
@@ -37,12 +35,9 @@ export default function TextImageBlock({
           imageLocation === 'LEFT' && displayStyle === '1x1',
       })}
     >
-      <motion.article
-        initial={fadeInText ? { y: 20, opacity: 0 } : { opacity: 1 }}
-        animate={fadeInText ? { y: 0, opacity: 1 } : { opacity: 1 }}
-        transition={
-          fadeInText && { duration: 0.5, delay: fadeInTextDelay ?? 0 }
-        }
+      <TextImageAnimation
+        fadeIn={fadeInText}
+        fadeInDelay={fadeInTextDelay}
         className='grid-text prose col-span-2 grow pr-4 dark:prose-invert prose-h1:mb-1 prose-a:text-pine-600 hover:prose-a:text-pine-700 dark:prose-a:text-gray-500 dark:hover:prose-a:text-gray-400'
       >
         <StructuredText
@@ -62,14 +57,12 @@ export default function TextImageBlock({
             }
           }}
         />
-      </motion.article>
-      {((image as FileField)?.responsiveImage as any) && (
-        <motion.div
-          initial={fadeInImage ? { y: 20, opacity: 0 } : { opacity: 1 }}
-          animate={fadeInImage ? { y: 0, opacity: 1 } : { opacity: 1 }}
-          transition={
-            fadeInImage && { duration: 0.5, delay: fadeInImageDelay ?? 0 }
-          }
+      </TextImageAnimation>
+
+      {image?.responsiveImage && (
+        <TextImageAnimation
+          fadeIn={fadeInImage}
+          fadeInDelay={fadeInImageDelay}
           className={cn('grid-image mx-auto my-4 md:mb-auto', {
             'col-span-2': displayStyle === '1x1',
             'col-span-1': displayStyle === '2x1',
@@ -84,16 +77,9 @@ export default function TextImageBlock({
               },
             )}
           >
-            {/* eslint-disable-next-line jsx-a11y/alt-text */}
-            <Image
-              data={(image as FileField)?.responsiveImage as any}
-              className={cn('translate-z-0', {
-                'rounded-full': imageStyle === 'Round',
-                'rounded-lg': imageStyle === 'Rounded Corners',
-              })}
-            />
+            <SRCImage data={image.responsiveImage} />
           </div>
-        </motion.div>
+        </TextImageAnimation>
       )}
     </div>
   );
