+++ src/components/Layouts/Footer/Footer.tsx
@@ -1,39 +1,21 @@
-'use client';
-
 import React from 'react';
 import Icon from 'components/Icons';
-import { motion } from 'framer-motion';
+
+import FooterEmojiAnimation from './FooterEmojiAnimation';
 
 export default function Footer() {
   return (
     <footer className='container mx-auto md:mb-10'>
       <div className='container mx-auto flex flex-col items-center space-y-4 bg-pine-300/80 pb-10 pt-12 font-extralight dark:border dark:border-gray-700/50 dark:bg-zinc-950/90 md:max-w-6xl md:rounded-sm'>
         <ul className='p-4'>
-          <motion.li
-            className='text-3xl'
-            animate={{ opacity: [1, 1, 0, 1, 1] }}
-            transition={{ duration: 3, repeat: Infinity }}
-            drag
-            dragConstraints={{
-              top: -500,
-              left: -150,
-              right: 150,
-              bottom: 50,
-            }}
-            dragTransition={{ bounceStiffness: 600, bounceDamping: 20 }}
-            dragElastic={0.5}
-            whileTap={{ cursor: 'grabbing' }}
-            whileDrag={{ opacity: 1 }}
-          >
-            ✌️
-          </motion.li>
+          <FooterEmojiAnimation />
         </ul>
         <ul className='flex p-4'>
           <li className='px-2'>
             <a
               href='https://www.linkedin.com/in/mnleth/'
               target='_blank'
-              rel='noopener norefferer noreferrer'
+              rel='noopener noreferrer'
               className='hover:text-pine-600'
             >
               <Icon symbol='linkedin' />
@@ -43,7 +25,7 @@ export default function Footer() {
             <a
               href='https://www.facebook.com/mark.nygaard'
               target='_blank'
-              rel='noopener norefferer noreferrer'
+              rel='noopener noreferrer'
               className='hover:text-pine-600'
             >
               <Icon symbol='facebook' />
@@ -53,7 +35,7 @@ export default function Footer() {
             <a
               href='https://github.com/MarkNygaard'
               target='_blank'
-              rel='noopener norefferer noreferrer'
+              rel='noopener noreferrer'
               className='hover:text-pine-600'
             >
               <Icon symbol='github' />
@@ -63,7 +45,7 @@ export default function Footer() {
             <a
               href='mailto:mark.nygaard@hotmail.com'
               target='_self'
-              rel='noopener norefferer'
+              rel='noopener noreferrer'
               className='hover:text-pine-600'
             >
               <Icon symbol='mail' />
