+++ src/components/Primitives/CardSpotlight.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import React, { useRef, useState } from 'react';
 import { motion } from 'framer-motion';
